AI-Powered Anime Recommendation System/Project Outputs/requirements.txt
langchain==1.2.1
langchain-core==1.2.6
langchain-community==0.4.1
langchain-chroma==1.1.0
langchain-groq==1.1.1
langchain-text-splitters==1.1.0
chromadb==1.4.0
sentence-transformers==5.2.0
langchain-huggingface==1.2.0
streamlit==1.52.2
pandas==2.3.3
python-dotenv==1.2.1
pillow==12.1.0
setuptools
